+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 See the [keepachangelog.com description](https://keepachangelog.com/en/1.0.0/).
 
-## 0.26.0 - 2025-06-16
+## 0.26.0 - 2025-06-26
 
 * Added
   * twins::findCoveringWidget() - now the WndManager::hide() redraw only neccessary part of the underlying window
+  * String::operator<()
 * Changed
   * WindowStateBase::invalidateImpl() updated - taken from demo, operates on invalidatedWgts